fix(anta.tests): Fix VerifyReachability failure messages
